backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.10.3
email-validator==2.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow>=9.0.0
supabase==2.3.4
transformers==4.36.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
